refactor(#1656): rename v5_compressed_legacy module to row_decoder
The v5_compressed_legacy module is NOT legacy — it is THE live row
decode engine. Rename the module directory and all module-path
references to row_decoder so the name stops lying.

Pure rename: no behavior/signature changes. The DataFormat::V5CompressedLegacy
enum variant (a format label) and V5CompressedLegacyParser type name are
intentionally left unchanged — only the module path is renamed.

Closes #1656
