+++ cqlite-core/src/storage/sstable/reader/parsing/row_decoder/block_emit.rs
